# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API App - Clinica",
        "version": "1.0"
    },
    "paths": {
...
        "/api/show/productos": {
            "get": {
                "tags": [
                    "alimentos"
                ],
                "summary": "Mostrando todos los productos",
                "operationId": "showProductos",
                "responses": {
                    "200": {
                        "description": "Showing al products"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "Product not registered yet"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/import/subcategorias": {
            "post": {
                "tags": [
                    "importFile"
                ],
                "summary": "Importar Sucategorias del excel",
                "operationId": "importSubcategorias",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file_excel": {
                                        "description": "file_excel",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategories imported"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/subcategorias": {
            "get": {
                "tags": [
                    "subcategoria-alimentos"
                ],
                "summary": "Mostrando subcategorias de Alimentos",
                "operationId": "showSubcategorias",
                "responses": {
                    "200": {
                        "description": "Showing subcategories of Alimentos"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/subcategoria/{subcategoria}": {
            "get": {
                "tags": [
                    "subcategoria-alimentos"
                ],
                "summary": "Mostrando la Subcategoria de Alimentos con su respectiva Categoria",
                "operationId": "showSubcategory",
                "parameters": [
                    {
                        "name": "subcategoria",
                        "in": "path",
                        "description": "Id of Subcategoria",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing subcategory of Alimentos with it respective Category"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/create/subcategoria": {
            "post": {
                "tags": [
                    "subcategoria-alimentos"
                ],
                "summary": "Crear subcategoria",
                "operationId": "createSubcategoria",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update/subcategoria/{subcategoria}": {
            "post": {
                "tags": [
                    "subcategoria-alimentos"
                ],
                "summary": "Actualizar subcategoria",
                "operationId": "updateSubcategoria",
                "parameters": [
                    {
                        "name": "subcategoria",
                        "in": "path",
                        "description": "Id of Subcategoria",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/subcategoria/{subcategoria}/productos": {
            "get": {
                "tags": [
                    "subcategoria-alimentos"
                ],
                "summary": "Mostrando productos de una subcategoria",
                "operationId": "showSubcategoryProducts",
                "parameters": [
                    {
                        "name": "subcategoria",
                        "in": "path",
                        "description": "Id of Subcategoria",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing products of a subcategory"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}